requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astroid==3.3.8
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.1
cycler==0.12.1
dill==0.3.9
Django==5.1.5
dropbox==12.0.2
fastapi==0.115.12
fonttools==4.56.0
h11==0.16.0
idna==3.10
iniconfig==2.0.0
isort==6.0.0
kiwisolver==1.4.8
matplotlib==3.10.0
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.2.2
packaging==24.2
pandas==2.3.0
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
pydantic==2.11.5
pydantic_core==2.33.2
pygame==2.6.1
pylint==3.3.4
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
stone==3.3.1
tomlkit==0.13.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
